+++ packages/models/lib/src/models/changelog/changelog_entry.dart
@@ -1,11 +1,12 @@
 import 'package:copy_with_extension/copy_with_extension.dart';
+import 'package:equatable/equatable.dart';
 import 'package:json_annotation/json_annotation.dart';
 
 part 'changelog_entry.g.dart';
 
 @CopyWith()
 @JsonSerializable()
-class ChangelogEntry {
+class ChangelogEntry extends Equatable {
   const ChangelogEntry({
     required this.date,
     required this.changes,
@@ -21,4 +22,10 @@ class ChangelogEntry {
 
   @override
   String toString() => 'ChangelogEntry(date: $date, changes: $changes)';
+
+  @override
+  List<Object?> get props => [
+    date,
+    changes,
+  ];
 }