+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # 283 - Clean up Code
 
 - Remove hashCode and == methods
+- Replace with `Equatable` package